python/requirements.txt
pyfiglet
pyserial